% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSSN.R
\name{plotSSN}
\alias{plotSSN}
\title{Plotting Function for SpatialStreamNetwork Objects}
\usage{
plotSSN(x, VariableName = NULL, addFuncColName = NULL, nbrks = 8,
  strCol = "blue", strLwd = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{\link[=SpatialStreamNetwork-class]{SpatialStreamNetwork}}.}

\item{VariableName}{a response variable name in the data.frame of observed data in the SpatialStreamNetwork object.}

\item{addFuncColName}{the additive function column name in the data.frame of lines data in the SpatialStreamNetwork object.  The size of the streams are scaled by this column. If NULL (default) all lines are are same size.}

\item{nbrks}{the number of classes for coloring the predictions (or standard errors) according to their value.  The default is 8.}

\item{strCol}{color of the stream lines. Default is 'blue'.}

\item{...}{other arguments passed from the points function}

\item{strLwd:}{line width for streams.  Ignored if addFuncColName is used.}
}
\value{
a map of a SpatialStreamNetwork object using sp plotting functions.
}
\description{
Plotting Function for SpatialStreamNetwork Objects
}
\author{
Jay Ver Hoef
}
\seealso{
\code{\link{plotPointsRGB}}, \code{\link{rect}},
}

