% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSNreachNei.R
\name{SSNreachNei}
\alias{SSNreachNei}
\title{Function for finding neighboring reaches in SpatialStreamNetwork Objects}
\usage{
SSNreachNei(object)
}
\arguments{
\item{object}{an object of class \code{\link[=SpatialStreamNetwork-class]{SpatialStreamNetwork}}.}
}
\value{
a sparse matrix with indicators of neighboring reaches.  
The rownames and column names are included, and match those from the 
rownames of the reaches in the \code{\link[=SpatialStreamNetwork-class]{SpatialStreamNetwork}} object.
}
\description{
Function for finding neighboring reaches in SpatialStreamNetwork Objects
}
\author{
Jay Ver Hoef
}
